fix: update import paths in example plugins after move
After moving example plugins to example-plugins/ subdirectory, the relative
imports needed to be updated from ./ to ../ to correctly reference files
in the parent js/ directory.

Fixes failing tests:
- test_extension_hooks.js
- test_feature_registry.js
