+++ src/canvas_chat/static/js/example-plugins/example-test-plugin.js
@@ -3,8 +3,8 @@
  * Demonstrates the FeaturePlugin API for testing and documentation
  */
 
-import { FeaturePlugin } from './feature-plugin.js';
-import { NodeType, createNode } from './graph-types.js';
+import { FeaturePlugin } from '../feature-plugin.js';
+import { NodeType, createNode } from '../graph-types.js';
 
 /**
  * SimpleTestPlugin is a minimal example feature plugin.
